refactor(backend): report real artist stats instead of hash-based mocks
작가 포트폴리오의 판매 수·평점·리뷰 수가 작가명 해시로 만들어낸 가짜 값이었고
지역도 "Seoul, Korea" 고정이었다.

- 판매 수는 실제 주문에서 세고, 지역은 같은 활동명 사용자 프로필의 시·도를 쓴다.
- 리뷰 도메인이 없으므로 평점·리뷰 수는 null 로 내려 화면이 '-' 로 표시하게 한다
  (없는 평점을 있는 것처럼 보여주지 않는다).

+++ backend/src/main/java/com/example/artnara/domain/artist/dto/ArtistDto.java
@@ -9,9 +9,10 @@ public class ArtistDto {
             String introduction,
             String location,
             int artworkCount,
-            int salesCount,
-            double rating,
-            int reviewCount,
+            long salesCount,
+            /** 리뷰 기능이 아직 없어 null (있는 척하지 않는다) */
+            Double rating,
+            Integer reviewCount,
             List<ArtworkSummary> artworks
     ) {}
